Keeps computation on appropriate devices.
Previously this would attempt to move the tokenizer to CUDA. This leads to a crash (see #13) if one requires a non-CUDA accelerator.

Furthermore, apparently there is no real gain for tokenizing on GPU anyways: https://stackoverflow.com/a/66116061

I have taken the liberty of doing a cheeky rename too; I think this is more to style.

Closes #13.