Sweep runs as subprocess
This modifies the sweep so that individual runs are all done in a
subprocess. The reason for this is a belief that the OS will do a better
job cleaning up memory (either on success or crash) than we can expect
to happen even if we explicitly call the GC for Python and/or CUDA.
We'll see if I'm right about this soon enough.