Indirectly adds metric-based testing
This PR first improves the prediction-based testing by very lightly
tuning the two models:

* We now use language-appropriate pretrained models
* We increase the learning rates substantially; the goal is to overfit
  here anyways
* We now do 10 epochs of training; this is not very expensive compared
  to the time spent downloading dependencies and pretrained models
* We actually reuse the best checkpoint (!)

This gives us, to rounding, perfect resubstitution accuracy, so any
model which gives the expected predictions has successfully overfit. To
confirm this, we run `udtube test` and the results appear in the logs.

This is an indirect, lightweight solution and closes #64.

Other details:

* I changed `assertFileExists` to `assertNonEmptyFileExists`, just to be
  sure. `os.stat` is very cheap.
* I moved some testdata files around.
* I added docs on how to regenerate the data if models ever change. This
  should make fixing this change-detector test much easier!