Checks for no enabled heads
If there are no heads enabled, the classifier module will now raise an
error (instead of crashing at a later stage with a less informative
error: it'll say something about the optimizer not knowing what to do
with a Torch module with no parameters---presumably the classifier).

Also adds a light test to make sure this fires.